refactor(provider): split provider file
