+++ .vscode/settings.json
@@ -1,0 +1,13 @@
+{
+    "files.exclude": {
+        "**/*.g.dart": true,
+        "**/*.mocks.dart": true,
+        "**/*.pickled.dart": true,
+        ".flutter-plugins": true,
+        ".flutter-plugins-dependencies": true,
+        ".dart_tool": true,
+        ".idea": true,
+        ".metadata": true,
+        "freenance.iml": true,
+    }
+}
